+++ app/Http/Requests/ShowTranslationRequest.php
@@ -33,4 +33,9 @@ class ShowTranslationRequest extends FormRequest
             'errors' => $validator->errors(),
         ], 422));
     }
+
+    public function getTranslationId(): int
+    {
+        return (int) $this->route('translation');
+    }
 }